[libc][posix] move libc.c/.h to posix folder
