+++ components/libc/compilers/armlibc/SConscript
@@ -1,18 +1,12 @@
 from building import *
 Import('rtconfig')
 
-src   = Glob('*.c') + Glob('*.cpp')
-cwd   = GetCurrentDir()
+src   = Glob('*.c')
 group = []
 
-CPPPATH = [cwd]
 CPPDEFINES = ['RT_USING_ARM_LIBC']
 
-if GetDepend('RT_USING_MODULE') == False:
-    SrcRemove(src, ['libc_syms.c'])
-
 if rtconfig.PLATFORM == 'armcc' or rtconfig.PLATFORM == 'armclang':
-    group = DefineGroup('libc', src, depend = ['RT_USING_LIBC'], 
-        CPPPATH = CPPPATH, CPPDEFINES = CPPDEFINES)
+    group = DefineGroup('libc', src, depend = ['RT_USING_LIBC'], CPPDEFINES = CPPDEFINES)
 
 Return('group')
