+++ components/libc/compilers/common/SConscript
@@ -8,13 +8,13 @@ group = []
 CPPPATH = [cwd]
 CPPDEFINES = []
 
+if rtconfig.CROSS_TOOL == 'keil':
+    CPPDEFINES += ['__CLK_TCK=RT_TICK_PER_SECOND']
+
 if GetDepend('RT_USING_LIBC'):
     src += Glob('*.c')
 elif GetDepend('RT_LIBC_USING_TIME'):
     src += ['time.c']
-
-if rtconfig.CROSS_TOOL == 'keil':
-    CPPDEFINES += ['__CLK_TCK=RT_TICK_PER_SECOND']
 
 group = DefineGroup('libc', src, depend = [], CPPPATH = CPPPATH, CPPDEFINES = CPPDEFINES)
