+++ components/libc/compilers/dlib/SConscript
@@ -1,25 +1,20 @@
 from building import *
-
 Import('rtconfig')
 
 src   = Glob('*.c')
-cwd   = GetCurrentDir()
 group = []
 
-CPPPATH    = [cwd]
 CPPDEFINES = ['RT_USING_DLIBC']
 
 if rtconfig.PLATFORM == 'iar':
-
     if GetDepend('RT_USING_POSIX'):
         from distutils.version import LooseVersion
         from iar import IARVersion
 
         CPPDEFINES = CPPDEFINES + ['_DLIB_FILE_DESCRIPTOR']
-
         if LooseVersion(IARVersion()) < LooseVersion("8.20.1"):
             CPPDEFINES = CPPDEFINES + ['_DLIB_THREAD_SUPPORT']
 
-    group = DefineGroup('libc', src, depend = ['RT_USING_LIBC'], CPPPATH = CPPPATH, CPPDEFINES = CPPDEFINES)
+    group = DefineGroup('libc', src, depend = ['RT_USING_LIBC'], CPPDEFINES = CPPDEFINES)
 
 Return('group')
